moved app.js
