+++ src/app/screens/Login/index.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import 'react-native-gesture-handler';
+import {
+  SafeAreaView,
+  StatusBar,
+  Text,
+  TouchableOpacity,
+  useColorScheme,
+} from 'react-native';
+import { Colors } from 'react-native/Libraries/NewAppScreen';
+
+import styles from './styles';
+
+const Login = () => {
+  const isDarkMode = useColorScheme() === 'dark';
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  const handlePress = () => console.warn("clicked");
+
+  return (
+    <SafeAreaView style={[styles.container, backgroundStyle]}>
+      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
+        <TouchableOpacity onPress={handlePress}>
+            <Text>Go to Home</Text>
+        </TouchableOpacity>
+    </SafeAreaView>
+  );
+};
+
+export default Login;
