+++ src/app/screens/Login/styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center'
+  }
+});
+
